Implement initial prototype
The skill responds to utterance of the form "play the file FILE NAME".
When invoked, the skill extracts the "FILE NAME" portion of the
utterance and uses to find a corresponding file in the `media/`
directory of the skill using Mycroft's built-in fuzzy search. It then
uses Mycroft's built-in audio service to play the file, if a match is
found.